+++ src/IDatabaseConnection.ts
@@ -1,0 +1,31 @@
+// Copyright (C) 2017  Norman Breau
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import { Readable } from 'stream';
+
+export interface IDatabaseConnection {
+    getInstantiationStack(): string;
+    getAPI(): any;
+    isReadOnly(): boolean;
+    setTimeout(timeout: number): void;
+    getTimeout(): number;
+    query(query: string, params?: any): Promise<any>;
+    stream(query: string, params?: any, streamOptions?: any): Readable;
+    close(forceClose: boolean): Promise<void>;
+    startTransaction(): Promise<void>;
+    isTransaction(): boolean;
+    commit(): Promise<void>;
+    rollback(): Promise<void>;
+}