+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.19.0 (July 17, 2019)
+- Added managed connections
+- Migrated from TSLint to ESLint
+
 # 0.18.4 (July 10, 2019)
 - made Logger.log protected
 - Ensure generated JWT tokens will be unique.